Switch to accurate conversion
This switches the implementation to a public-domain implementation
designed by "x4u". This fixes any and all accuracy problems related
to subnormals.

See: https://stackoverflow.com/questions/6162651/half-precision-floating-point-in-java
See: README-algo.txt

This also bumps the project version to 1.0.4-SNAPSHOT and, out of
necessity, uses the 1.1.0 version of the primogenitor POM.

Fix #1